rag-medical-chatbot/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
duckduckgo-search
groq
pypdf
